Refactor JS modules
