+++ index.html
@@ -18,5 +18,5 @@
 		    </div>
 		</div>
 	</body>
-    <script src="src/js/index.js"></script>
+    <script type="module" src="src/js/index.js"></script>
 </html>
