+++ src/html/render_filedrop.html
@@ -46,8 +46,6 @@
         <canvas class="chart-canvas" id="chart-canvas"></canvas>
         <canvas class="chart-canvas" id="chart-canvas-er"></canvas>
     </div>
-    <script src="../js/rift_essentials.js"></script>
-    <script src="../js/render.js"></script>
-    <script src="../js/onload.js"></script>
+    <script type="module" src="../js/onload.js"></script>
 </body>
 </html>
