+++ src/js/rift_essentials_bin.js
