+++ tests/test.html
@@ -5,9 +5,9 @@
     <title>Test</title>
 </head>
 <body>
-    <script src="../src/js/rift_essentials.js"></script>
-    <script src="../src/js/render.js"></script>
-    <script src="../src/js/decode_bin.js"></script>
-    <script src="../src/js/test_compare_bin_json.js"></script>
+    <script type="module" src="../src/js/rift_essentials_json.js"></script>
+    <script type="module" src="../src/js/render.js"></script>
+    <script type="module" src="../src/js/rift_essentials_bin.js"></script>
+    <script type="module" src="../src/js/test_compare_bin_json.js"></script>
 </body>
 </html>
